Render certificates of completion for a participant in a cohort

Render certificates of completion for a participant in a cohort

Usage

create_certificate(
  cohort_name,
  first_name,
  last_name,
  start_date,
  end_date,
  cohort_website,
  output_dir = "."
)

Arguments

cohort_name

The name of the cohort

first_name

First name of participant

last_name

Last name of participant

start_date

cohort start date

end_date

cohort end date

cohort_website

cohort website

output_dir

output directory for certificates. Default "."

Value

Saves the file to your current working directory, and returns the path to the file

Examples

if (FALSE) { # \dontrun{
create_certificate(cohort_name = "2023-fred-hutch",
                   participant_name = "Name",
                   start_date = "Sep 19",
                   end_date = "Oct 19",
                   cohort_website = "https://openscapes.github.io/2023-fred-hutch/")
} # }